src/django/requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
attrs==22.2.0
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==3.2.0
cond-rnn==3.2.2
distlib==0.3.6
Django==4.0.10
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.0
filelock==3.9.0
flatbuffers==1.12
gast==0.4.0
geographiclib==2.0
geopy==2.4.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.57.0
gunicorn==20.1.0
h5py==3.9.0
idna==3.4
inflection==0.5.1
install==1.3.5
joblib==1.3.2
jsonschema==4.17.3
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
pipenv==2023.2.18
platformdirs==3.0.0
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.6.0
pymssql==2.2.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sqlparse==0.4.3
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.0
tensorflow-addons==0.21.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
typeguard==2.13.3
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.19.0
virtualenv-clone==0.5.7
Werkzeug==2.3.7
wrapt==1.15.0
